(#36) fix: cookie 커스텀훅이 아닌 util 함수로 관리
